feat: populate deploy-config.json with env variables

+++ prepare.sh
@@ -9,6 +9,17 @@ set -e
 # Setting an environment variable for deployment
 export IMPL_SALT=$(openssl rand -hex 32)
 
+# Check if both L1_BLOCKHASH and L1_TIMESTAMP are set or unset
+if [ -n "$L1_BLOCKHASH" ] && [ -z "$L1_TIMESTAMP" ] || [ -z "$L1_BLOCKHASH" ] && [ -n "$L1_TIMESTAMP" ]; then
+  echo "Error: Both L1_BLOCKHASH and L1_TIMESTAMP must be set or unset."
+  exit 1
+elif [ -z "$L1_BLOCKHASH" ] && [ -z "$L1_TIMESTAMP" ]; then
+  # Fetch block details if both variables are unset
+  block=$(cast block finalized --rpc-url $L1_RPC_URL)
+  export L1_TIMESTAMP=$(echo "$block" | awk '/timestamp/ { print $2 }')
+  export L1_BLOCKHASH=$(echo "$block" | awk '/hash/ { print $2 }')
+fi
+
 # Build the Optimism Monorepo
 cd /app/data/optimism
 pnpm install
@@ -19,13 +30,16 @@ pnpm build
 cd /app/data/op-geth
 make geth
 
-# Configure network
-cd /app/data/optimism/packages/contracts-bedrock
-./scripts/getting-started/config.sh
-cp -f ./deploy-config/$DEPLOYMENT_CONTEXT.json /app/data/configurations/deploy-config.json
-
-# Update the configuration with custom values
-/app/update-config.sh
+# Check if deploy-config.json exists
+if [ -f "/app/deploy-config.json" ]; then
+  # Populate deploy-config.json with env variables
+  envsubst < /app/deploy-config.json > /app/temp-deploy-config.json && mv /app/temp-deploy-config.json /app/data/configurations/deploy-config.json
+else
+  # If deploy-config.json does not exist, use config.sh to generate it
+  cd /app/data/optimism/packages/contracts-bedrock
+  ./scripts/getting-started/config.sh
+  cp -f ./deploy-config/$DEPLOYMENT_CONTEXT.json /app/data/configurations/deploy-config.json
+fi
 
 # Deploy the L1 contracts
 forge script scripts/Deploy.s.sol:Deploy --private-key $GS_ADMIN_PRIVATE_KEY --broadcast --rpc-url $L1_RPC_URL
